test(e2e): share variant action helpers between specs
Move create/delete variant helpers into e2e/tests/variants/utils.ts so Vision
and the variants tool specs can reuse the same API actions, including deleting
a published variant document before its definition.
