+++ e2e/tests/vision/vision.spec.ts
@@ -12,6 +12,8 @@ import {
   createVariantDefinition,
   createVariantDocument,
   deleteVariantDefinition,
+} from '../variants/utils'
+import {
   encodeQueryString,
   fetchPublishedVariantOverlay,
   getVisionRegions,
